requirements.txt
python_version >= "3.10"
matplotlib == 3.7.2
gymnasium == 0.28.1
pandas == 2.0.3
stable-baselines3 == 2.0.0a13
sb3-contrib == 2.0.0a13
tensorboard == 2.13.0
tqdm == 4.65.1
numpy == 1.25.2
rich == 13.5.2
torch == 2.0.1
pyyaml == 6.0.2
protobuf == 4.24.0